Move tests for testing code to testing package
