+++ sympy/testing/tests/diagnose_imports.py
