+++ sympy/testing/tests/test_code_quality.py
@@ -9,7 +9,7 @@ import ast
 
 from sympy.core.compatibility import PY3
 from sympy.testing.pytest import raises
-from sympy.utilities.quality_unicode import test_this_file_encoding
+from sympy.testing.quality_unicode import test_this_file_encoding
 
 # System path separator (usually slash or backslash) to be
 # used with excluded files, e.g.
