+++ sympy/testing/tests/test_module_imports.py
