+++ sympy/testing/tests/test_pytest.py
