+++ sympy/utilities/tests/test_autowrap.py
@@ -13,7 +13,7 @@ from sympy.utilities.codegen import (
     CCodeGen, C99CodeGen, CodeGenArgumentListError, make_routine
 )
 from sympy.testing.pytest import raises
-from sympy.utilities.tmpfiles import TmpFileManager
+from sympy.testing.tmpfiles import TmpFileManager
 
 
 def get_string(dump_fn, routines, prefix="file", **kwargs):
